resonanceFit: move configuration file reading to separate file
Move the reading of the YAML file to a separate source file.
